healthcheck/requirements.txt
Flask
flask-mail
python-dotenv
flask_apscheduler